requirements.txt
fastapi==0.115.0
fastapi[standard]==0.115.0
pydantic==2.9.2
feedgen==1.0.0
feedparser==6.0.11
loguru==0.7.2
openai==1.51.2
peewee==3.17.6
python-dotenv==1.0.1
qdrant_client==1.12.0
trafilatura==1.12.2
uvicorn==0.31.0
psycopg2-binary==2.9.9
pyyaml==6.0.2
schedule==1.2.2
ruff==0.6.9
isort==6.0.0b2
black==24.10.0
requests==2.32.3
scalar_fastapi==1.0.3
redis==4.6.0
fastapi-cache2[redis]==0.2.2